requirements.txt
openai-whisper
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
pyaudio==0.2.13
pynput==1.7.6
pywin32==306
pystray==0.19.5
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
sounddevice==0.4.6
soundfile==0.12.1
keyboard==0.13.5
pyperclip==1.8.2
sentencepiece>=0.1.99
sacremoses
protobuf>=3.20.0
customtkinter==5.2.1
numba
more-itertools
tiktoken
regex
tqdm
ffmpeg-python